refactored all statistic module and modules structure
