+++ frontend/js/modules/statisticsModule/statisticsModule.html
